+++ internal/impl/prometheus/metrics_prometheus.go
@@ -100,7 +100,8 @@ func (p *promTiming) Timing(val int64) {
 //------------------------------------------------------------------------------
 
 type promCounterVec struct {
-	ctr *prometheus.CounterVec
+	ctr   *prometheus.CounterVec
+	count int
 }
 
 func (p *promCounterVec) With(labelValues ...string) metrics.StatCounter {
@@ -110,7 +111,8 @@ func (p *promCounterVec) With(labelValues ...string) metrics.StatCounter {
 }
 
 type promTimingVec struct {
-	sum *prometheus.SummaryVec
+	sum   *prometheus.SummaryVec
+	count int
 }
 
 func (p *promTimingVec) With(labelValues ...string) metrics.StatTimer {
@@ -120,7 +122,8 @@ func (p *promTimingVec) With(labelValues ...string) metrics.StatTimer {
 }
 
 type promTimingHistVec struct {
-	sum *prometheus.HistogramVec
+	sum   *prometheus.HistogramVec
+	count int
 }
 
 func (p *promTimingHistVec) With(labelValues ...string) metrics.StatTimer {
@@ -131,7 +134,8 @@ func (p *promTimingHistVec) With(labelValues ...string) metrics.StatTimer {
 }
 
 type promGaugeVec struct {
-	ctr *prometheus.GaugeVec
+	ctr   *prometheus.GaugeVec
+	count int
 }
 
 func (p *promGaugeVec) With(labelValues ...string) metrics.StatGauge {
@@ -155,10 +159,10 @@ type prometheusMetrics struct {
 	pusher *push.Pusher
 	reg    *prometheus.Registry
 
-	counters   map[string]*prometheus.CounterVec
-	gauges     map[string]*prometheus.GaugeVec
-	timers     map[string]*prometheus.SummaryVec
-	timersHist map[string]*prometheus.HistogramVec
+	counters   map[string]*promCounterVec
+	gauges     map[string]*promGaugeVec
+	timers     map[string]*promTimingVec
+	timersHist map[string]*promTimingHistVec
 
 	mut sync.Mutex
 }
@@ -172,10 +176,10 @@ func newPrometheus(config metrics.Config, log log.Modular) (metrics.Type, error)
 		useHistogramTiming: promConf.UseHistogramTiming,
 		histogramBuckets:   promConf.HistogramBuckets,
 		reg:                prometheus.NewRegistry(),
-		counters:           map[string]*prometheus.CounterVec{},
-		gauges:             map[string]*prometheus.GaugeVec{},
-		timers:             map[string]*prometheus.SummaryVec{},
-		timersHist:         map[string]*prometheus.HistogramVec{},
+		counters:           map[string]*promCounterVec{},
+		gauges:             map[string]*promGaugeVec{},
+		timers:             map[string]*promTimingVec{},
+		timersHist:         map[string]*promTimingHistVec{},
 	}
 
 	if len(p.histogramBuckets) == 0 {
@@ -247,23 +251,30 @@ func (p *prometheusMetrics) GetCounterVec(path string, labelNames ...string) met
 		})
 	}
 
-	var ctr *prometheus.CounterVec
+	var pv *promCounterVec
 
 	p.mut.Lock()
 	var exists bool
-	if ctr, exists = p.counters[path]; !exists {
-		ctr = prometheus.NewCounterVec(prometheus.CounterOpts{
+	if pv, exists = p.counters[path]; !exists {
+		ctr := prometheus.NewCounterVec(prometheus.CounterOpts{
 			Name: path,
 			Help: "Benthos Counter metric",
 		}, labelNames)
 		p.reg.MustRegister(ctr)
-		p.counters[path] = ctr
+
+		pv = &promCounterVec{
+			ctr:   ctr,
+			count: len(labelNames),
+		}
+		p.counters[path] = pv
 	}
 	p.mut.Unlock()
 
-	return &promCounterVec{
-		ctr: ctr,
+	if pv.count != len(labelNames) {
+		p.log.Errorf("Metrics label mismatch %v versus %v %v for name '%v', skipping metric", pv.count, len(labelNames), labelNames, path)
+		return metrics.Noop().GetCounterVec(path, labelNames...)
 	}
+	return pv
 }
 
 func (p *prometheusMetrics) GetTimer(path string) metrics.StatTimer {
@@ -282,45 +293,59 @@ func (p *prometheusMetrics) GetTimerVec(path string, labelNames ...string) metri
 		return p.getTimerHistVec(path, labelNames...)
 	}
 
-	var tmr *prometheus.SummaryVec
+	var pv *promTimingVec
 
 	p.mut.Lock()
 	var exists bool
-	if tmr, exists = p.timers[path]; !exists {
-		tmr = prometheus.NewSummaryVec(prometheus.SummaryOpts{
+	if pv, exists = p.timers[path]; !exists {
+		tmr := prometheus.NewSummaryVec(prometheus.SummaryOpts{
 			Name:       path,
 			Help:       "Benthos Timing metric",
 			Objectives: map[float64]float64{0.5: 0.05, 0.9: 0.01, 0.99: 0.001},
 		}, labelNames)
 		p.reg.MustRegister(tmr)
-		p.timers[path] = tmr
+
+		pv = &promTimingVec{
+			sum:   tmr,
+			count: len(labelNames),
+		}
+		p.timers[path] = pv
 	}
 	p.mut.Unlock()
 
-	return &promTimingVec{
-		sum: tmr,
+	if pv.count != len(labelNames) {
+		p.log.Errorf("Metrics label mismatch %v versus %v %v for name '%v', skipping metric", pv.count, len(labelNames), labelNames, path)
+		return metrics.Noop().GetTimerVec(path, labelNames...)
 	}
+	return pv
 }
 
 func (p *prometheusMetrics) getTimerHistVec(path string, labelNames ...string) metrics.StatTimerVec {
-	var tmr *prometheus.HistogramVec
+	var pv *promTimingHistVec
 
 	p.mut.Lock()
 	var exists bool
-	if tmr, exists = p.timersHist[path]; !exists {
-		tmr = prometheus.NewHistogramVec(prometheus.HistogramOpts{
+	if pv, exists = p.timersHist[path]; !exists {
+		tmr := prometheus.NewHistogramVec(prometheus.HistogramOpts{
 			Name:    path,
 			Help:    "Benthos Timing metric",
 			Buckets: p.histogramBuckets,
 		}, labelNames)
 		p.reg.MustRegister(tmr)
-		p.timersHist[path] = tmr
+
+		pv = &promTimingHistVec{
+			sum:   tmr,
+			count: len(labelNames),
+		}
+		p.timersHist[path] = pv
 	}
 	p.mut.Unlock()
 
-	return &promTimingHistVec{
-		sum: tmr,
+	if pv.count != len(labelNames) {
+		p.log.Errorf("Metrics label mismatch %v versus %v %v for name '%v', skipping metric", pv.count, len(labelNames), labelNames, path)
+		return metrics.Noop().GetTimerVec(path, labelNames...)
 	}
+	return pv
 }
 
 func (p *prometheusMetrics) GetGauge(path string) metrics.StatGauge {
@@ -335,23 +360,30 @@ func (p *prometheusMetrics) GetGaugeVec(path string, labelNames ...string) metri
 		})
 	}
 
-	var ctr *prometheus.GaugeVec
+	var pv *promGaugeVec
 
 	p.mut.Lock()
 	var exists bool
-	if ctr, exists = p.gauges[path]; !exists {
-		ctr = prometheus.NewGaugeVec(prometheus.GaugeOpts{
+	if pv, exists = p.gauges[path]; !exists {
+		ctr := prometheus.NewGaugeVec(prometheus.GaugeOpts{
 			Name: path,
 			Help: "Benthos Gauge metric",
 		}, labelNames)
 		p.reg.MustRegister(ctr)
-		p.gauges[path] = ctr
+
+		pv = &promGaugeVec{
+			ctr:   ctr,
+			count: len(labelNames),
+		}
+		p.gauges[path] = pv
 	}
 	p.mut.Unlock()
 
-	return &promGaugeVec{
-		ctr: ctr,
+	if pv.count != len(labelNames) {
+		p.log.Errorf("Metrics label mismatch %v versus %v %v for name '%v', skipping metric", pv.count, len(labelNames), labelNames, path)
+		return metrics.Noop().GetGaugeVec(path, labelNames...)
 	}
+	return pv
 }
 
 func (p *prometheusMetrics) Close() error {