+++ internal/integration/stream_test_helpers.go
@@ -454,7 +454,7 @@ func initInput(t testing.TB, env *streamTestEnvironment) iinput.Streamed {
 	assert.Empty(t, lints)
 
 	if env.mgr == nil {
-		env.mgr, err = manager.NewV2(s.ResourceConfig, nil, env.log, env.stats)
+		env.mgr, err = manager.New(s.ResourceConfig, nil, env.log, env.stats)
 		require.NoError(t, err)
 	}
 
@@ -483,7 +483,7 @@ func initOutput(t testing.TB, trans <-chan message.Transaction, env *streamTestE
 	assert.Empty(t, lints)
 
 	if env.mgr == nil {
-		env.mgr, err = manager.NewV2(s.ResourceConfig, nil, env.log, env.stats)
+		env.mgr, err = manager.New(s.ResourceConfig, nil, env.log, env.stats)
 		require.NoError(t, err)
 	}