+++ internal/manager/type_test.go
@@ -50,7 +50,7 @@ func TestManagerProcessorLabels(t *testing.T) {
 		conf.Bloblang = "root = this"
 		conf.Label = l
 
-		mgr, err := manager.NewV2(manager.NewResourceConfig(), nil, log.Noop(), noopStats())
+		mgr, err := manager.New(manager.NewResourceConfig(), nil, log.Noop(), noopStats())
 		require.NoError(t, err)
 
 		_, err = mgr.NewProcessor(conf)
@@ -69,7 +69,7 @@ func TestManagerProcessorLabels(t *testing.T) {
 		conf.Bloblang = "root = this"
 		conf.Label = l
 
-		mgr, err := manager.NewV2(manager.NewResourceConfig(), nil, log.Noop(), noopStats())
+		mgr, err := manager.New(manager.NewResourceConfig(), nil, log.Noop(), noopStats())
 		require.NoError(t, err)
 
 		_, err = mgr.NewProcessor(conf)
@@ -90,7 +90,7 @@ func TestManagerCache(t *testing.T) {
 	barCache.Label = "bar"
 	conf.ResourceCaches = append(conf.ResourceCaches, barCache)
 
-	mgr, err := manager.NewV2(conf, nil, testLog, noopStats())
+	mgr, err := manager.New(conf, nil, testLog, noopStats())
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -110,7 +110,7 @@ func TestManagerCacheList(t *testing.T) {
 	conf := manager.NewResourceConfig()
 	conf.ResourceCaches = append(conf.ResourceCaches, cacheFoo, cacheBar)
 
-	mgr, err := manager.NewV2(conf, nil, log.Noop(), noopStats())
+	mgr, err := manager.New(conf, nil, log.Noop(), noopStats())
 	require.NoError(t, err)
 
 	err = mgr.AccessCache(context.Background(), "foo", func(cache.V1) {})
@@ -133,14 +133,14 @@ func TestManagerCacheListErrors(t *testing.T) {
 	conf := manager.NewResourceConfig()
 	conf.ResourceCaches = append(conf.ResourceCaches, cFoo, cBar)
 
-	_, err := manager.NewV2(conf, nil, log.Noop(), noopStats())
+	_, err := manager.New(conf, nil, log.Noop(), noopStats())
 	require.EqualError(t, err, "cache resource label 'foo' collides with a previously defined resource")
 
 	cEmpty := cache.NewConfig()
 	conf = manager.NewResourceConfig()
 	conf.ResourceCaches = append(conf.ResourceCaches, cEmpty)
 
-	_, err = manager.NewV2(conf, nil, log.Noop(), noopStats())
+	_, err = manager.New(conf, nil, log.Noop(), noopStats())
 	require.EqualError(t, err, "cache resource has an empty label")
 }
 
@@ -154,7 +154,7 @@ func TestManagerBadCache(t *testing.T) {
 	badConf.Type = "notexist"
 	conf.ResourceCaches = append(conf.ResourceCaches, badConf)
 
-	if _, err := manager.NewV2(conf, nil, testLog, noopStats()); err == nil {
+	if _, err := manager.New(conf, nil, testLog, noopStats()); err == nil {
 		t.Fatal("Expected error from bad cache")
 	}
 }
@@ -170,7 +170,7 @@ func TestManagerRateLimit(t *testing.T) {
 	barRL.Label = "bar"
 	conf.ResourceRateLimits = append(conf.ResourceRateLimits, barRL)
 
-	mgr, err := manager.NewV2(conf, nil, log.Noop(), noopStats())
+	mgr, err := manager.New(conf, nil, log.Noop(), noopStats())
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -190,7 +190,7 @@ func TestManagerRateLimitList(t *testing.T) {
 	conf := manager.NewResourceConfig()
 	conf.ResourceRateLimits = append(conf.ResourceRateLimits, cFoo, cBar)
 
-	mgr, err := manager.NewV2(conf, nil, log.Noop(), noopStats())
+	mgr, err := manager.New(conf, nil, log.Noop(), noopStats())
 	require.NoError(t, err)
 
 	err = mgr.AccessRateLimit(context.Background(), "foo", func(ratelimit.V1) {})
@@ -213,14 +213,14 @@ func TestManagerRateLimitListErrors(t *testing.T) {
 	conf := manager.NewResourceConfig()
 	conf.ResourceRateLimits = append(conf.ResourceRateLimits, cFoo, cBar)
 
-	_, err := manager.NewV2(conf, nil, log.Noop(), noopStats())
+	_, err := manager.New(conf, nil, log.Noop(), noopStats())
 	require.EqualError(t, err, "rate limit resource label 'foo' collides with a previously defined resource")
 
 	cEmpty := ratelimit.NewConfig()
 	conf = manager.NewResourceConfig()
 	conf.ResourceRateLimits = append(conf.ResourceRateLimits, cEmpty)
 
-	_, err = manager.NewV2(conf, nil, log.Noop(), noopStats())
+	_, err = manager.New(conf, nil, log.Noop(), noopStats())
 	require.EqualError(t, err, "rate limit resource has an empty label")
 }
 
@@ -231,7 +231,7 @@ func TestManagerBadRateLimit(t *testing.T) {
 	badConf.Label = "bad"
 	conf.ResourceRateLimits = append(conf.ResourceRateLimits, badConf)
 
-	if _, err := manager.NewV2(conf, nil, log.Noop(), noopStats()); err == nil {
+	if _, err := manager.New(conf, nil, log.Noop(), noopStats()); err == nil {
 		t.Fatal("Expected error from bad rate limit")
 	}
 }
@@ -247,7 +247,7 @@ func TestManagerProcessor(t *testing.T) {
 	barProc.Label = "bar"
 	conf.ResourceProcessors = append(conf.ResourceProcessors, barProc)
 
-	mgr, err := manager.NewV2(conf, nil, log.Noop(), noopStats())
+	mgr, err := manager.New(conf, nil, log.Noop(), noopStats())
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -267,7 +267,7 @@ func TestManagerProcessorList(t *testing.T) {
 	conf := manager.NewResourceConfig()
 	conf.ResourceProcessors = append(conf.ResourceProcessors, cFoo, cBar)
 
-	mgr, err := manager.NewV2(conf, nil, log.Noop(), noopStats())
+	mgr, err := manager.New(conf, nil, log.Noop(), noopStats())
 	require.NoError(t, err)
 
 	err = mgr.AccessProcessor(context.Background(), "foo", func(iprocessor.V1) {})
@@ -290,14 +290,14 @@ func TestManagerProcessorListErrors(t *testing.T) {
 	conf := manager.NewResourceConfig()
 	conf.ResourceProcessors = append(conf.ResourceProcessors, cFoo, cBar)
 
-	_, err := manager.NewV2(conf, nil, log.Noop(), noopStats())
+	_, err := manager.New(conf, nil, log.Noop(), noopStats())
 	require.EqualError(t, err, "processor resource label 'foo' collides with a previously defined resource")
 
 	cEmpty := processor.NewConfig()
 	conf = manager.NewResourceConfig()
 	conf.ResourceProcessors = append(conf.ResourceProcessors, cEmpty)
 
-	_, err = manager.NewV2(conf, nil, log.Noop(), noopStats())
+	_, err = manager.New(conf, nil, log.Noop(), noopStats())
 	require.EqualError(t, err, "processor resource has an empty label")
 }
 
@@ -313,7 +313,7 @@ func TestManagerInputList(t *testing.T) {
 	conf := manager.NewResourceConfig()
 	conf.ResourceInputs = append(conf.ResourceInputs, cFoo, cBar)
 
-	mgr, err := manager.NewV2(conf, nil, log.Noop(), noopStats())
+	mgr, err := manager.New(conf, nil, log.Noop(), noopStats())
 	require.NoError(t, err)
 
 	err = mgr.AccessInput(context.Background(), "foo", func(i iinput.Streamed) {})
@@ -336,14 +336,14 @@ func TestManagerInputListErrors(t *testing.T) {
 	conf := manager.NewResourceConfig()
 	conf.ResourceInputs = append(conf.ResourceInputs, cFoo, cBar)
 
-	_, err := manager.NewV2(conf, nil, log.Noop(), noopStats())
+	_, err := manager.New(conf, nil, log.Noop(), noopStats())
 	require.EqualError(t, err, "input resource label 'foo' collides with a previously defined resource")
 
 	cEmpty := input.NewConfig()
 	conf = manager.NewResourceConfig()
 	conf.ResourceInputs = append(conf.ResourceInputs, cEmpty)
 
-	_, err = manager.NewV2(conf, nil, log.Noop(), noopStats())
+	_, err = manager.New(conf, nil, log.Noop(), noopStats())
 	require.EqualError(t, err, "input resource has an empty label")
 }
 
@@ -359,7 +359,7 @@ func TestManagerOutputList(t *testing.T) {
 	conf := manager.NewResourceConfig()
 	conf.ResourceOutputs = append(conf.ResourceOutputs, cFoo, cBar)
 
-	mgr, err := manager.NewV2(conf, nil, log.Noop(), noopStats())
+	mgr, err := manager.New(conf, nil, log.Noop(), noopStats())
 	require.NoError(t, err)
 
 	err = mgr.AccessOutput(context.Background(), "foo", func(ow ioutput.Sync) {})
@@ -382,20 +382,20 @@ func TestManagerOutputListErrors(t *testing.T) {
 	conf := manager.NewResourceConfig()
 	conf.ResourceOutputs = append(conf.ResourceOutputs, cFoo, cBar)
 
-	_, err := manager.NewV2(conf, nil, log.Noop(), noopStats())
+	_, err := manager.New(conf, nil, log.Noop(), noopStats())
 	require.EqualError(t, err, "output resource label 'foo' collides with a previously defined resource")
 
 	cEmpty := output.NewConfig()
 	conf = manager.NewResourceConfig()
 	conf.ResourceOutputs = append(conf.ResourceOutputs, cEmpty)
 
-	_, err = manager.NewV2(conf, nil, log.Noop(), noopStats())
+	_, err = manager.New(conf, nil, log.Noop(), noopStats())
 	require.EqualError(t, err, "output resource has an empty label")
 }
 
 func TestManagerPipeErrors(t *testing.T) {
 	conf := manager.NewResourceConfig()
-	mgr, err := manager.NewV2(conf, nil, log.Noop(), noopStats())
+	mgr, err := manager.New(conf, nil, log.Noop(), noopStats())
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -407,7 +407,7 @@ func TestManagerPipeErrors(t *testing.T) {
 
 func TestManagerPipeGetSet(t *testing.T) {
 	conf := manager.NewResourceConfig()
-	mgr, err := manager.NewV2(conf, nil, log.Noop(), noopStats())
+	mgr, err := manager.New(conf, nil, log.Noop(), noopStats())
 	if err != nil {
 		t.Fatal(err)
 	}