+++ internal/old/input/http_server_test.go
@@ -62,7 +62,7 @@ func TestHTTPBasic(t *testing.T) {
 	nTestLoops := 100
 
 	reg := apiRegGorillaMutWrapper{mut: mux.NewRouter()}
-	mgr, err := manager.NewV2(manager.NewResourceConfig(), reg, log.Noop(), metrics.Noop())
+	mgr, err := manager.New(manager.NewResourceConfig(), reg, log.Noop(), metrics.Noop())
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -243,7 +243,7 @@ func TestHTTPServerLifecycle(t *testing.T) {
 		_ = apiImpl.Shutdown(context.Background())
 	}()
 
-	mgr, err := manager.NewV2(manager.NewResourceConfig(), apiImpl, log.Noop(), metrics.Noop())
+	mgr, err := manager.New(manager.NewResourceConfig(), apiImpl, log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	conf := input.NewConfig()
@@ -307,7 +307,7 @@ func TestHTTPServerMetadata(t *testing.T) {
 	defer done()
 
 	reg := apiRegGorillaMutWrapper{mut: mux.NewRouter()}
-	mgr, err := manager.NewV2(manager.NewResourceConfig(), reg, log.Noop(), metrics.Noop())
+	mgr, err := manager.New(manager.NewResourceConfig(), reg, log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	conf := input.NewConfig()
@@ -369,7 +369,7 @@ func TestHTTPtServerPathParameters(t *testing.T) {
 	defer done()
 
 	reg := apiRegGorillaMutWrapper{mut: mux.NewRouter()}
-	mgr, err := manager.NewV2(manager.NewResourceConfig(), reg, log.Noop(), metrics.Noop())
+	mgr, err := manager.New(manager.NewResourceConfig(), reg, log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	conf := input.NewConfig()
@@ -433,7 +433,7 @@ func TestHTTPBadRequests(t *testing.T) {
 	t.Parallel()
 
 	reg := apiRegGorillaMutWrapper{mut: mux.NewRouter()}
-	mgr, err := manager.NewV2(manager.NewResourceConfig(), reg, log.Noop(), metrics.Noop())
+	mgr, err := manager.New(manager.NewResourceConfig(), reg, log.Noop(), metrics.Noop())
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -468,7 +468,7 @@ func TestHTTPTimeout(t *testing.T) {
 	t.Parallel()
 
 	reg := apiRegGorillaMutWrapper{mut: mux.NewRouter()}
-	mgr, err := manager.NewV2(manager.NewResourceConfig(), reg, log.Noop(), metrics.Noop())
+	mgr, err := manager.New(manager.NewResourceConfig(), reg, log.Noop(), metrics.Noop())
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -521,7 +521,7 @@ rate_limit_resources:
       interval: 60s
 `), &mgrConf))
 
-	mgr, err := manager.NewV2(mgrConf, reg, log.Noop(), metrics.Noop())
+	mgr, err := manager.New(mgrConf, reg, log.Noop(), metrics.Noop())
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -587,7 +587,7 @@ func TestHTTPServerWebsockets(t *testing.T) {
 
 	reg := apiRegGorillaMutWrapper{mut: mux.NewRouter()}
 
-	mgr, err := manager.NewV2(manager.NewResourceConfig(), reg, log.Noop(), metrics.Noop())
+	mgr, err := manager.New(manager.NewResourceConfig(), reg, log.Noop(), metrics.Noop())
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -681,7 +681,7 @@ rate_limit_resources:
       interval: 60s
 `), &mgrConf))
 
-	mgr, err := manager.NewV2(mgrConf, reg, log.Noop(), metrics.Noop())
+	mgr, err := manager.New(mgrConf, reg, log.Noop(), metrics.Noop())
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -761,7 +761,7 @@ func TestHTTPSyncResponseHeaders(t *testing.T) {
 	t.Parallel()
 
 	reg := apiRegGorillaMutWrapper{mut: mux.NewRouter()}
-	mgr, err := manager.NewV2(manager.NewResourceConfig(), reg, log.Noop(), metrics.Noop())
+	mgr, err := manager.New(manager.NewResourceConfig(), reg, log.Noop(), metrics.Noop())
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -909,7 +909,7 @@ func TestHTTPSyncResponseMultipart(t *testing.T) {
 	t.Parallel()
 
 	reg := apiRegGorillaMutWrapper{mut: mux.NewRouter()}
-	mgr, err := manager.NewV2(manager.NewResourceConfig(), reg, log.Noop(), metrics.Noop())
+	mgr, err := manager.New(manager.NewResourceConfig(), reg, log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	conf := input.NewConfig()
@@ -981,7 +981,7 @@ func TestHTTPSyncResponseHeadersStatus(t *testing.T) {
 	t.Parallel()
 
 	reg := apiRegGorillaMutWrapper{mut: mux.NewRouter()}
-	mgr, err := manager.NewV2(manager.NewResourceConfig(), reg, log.Noop(), metrics.Noop())
+	mgr, err := manager.New(manager.NewResourceConfig(), reg, log.Noop(), metrics.Noop())
 	if err != nil {
 		t.Fatal(err)
 	}