+++ internal/stream/manager/api_test.go
@@ -125,7 +125,7 @@ func (f *endpointReg) RegisterEndpoint(path, desc string, h http.HandlerFunc) {
 
 func TestTypeAPIDisabled(t *testing.T) {
 	r := &endpointReg{endpoints: map[string]http.HandlerFunc{}}
-	rMgr, err := bmanager.NewV2(bmanager.NewResourceConfig(), r, log.Noop(), metrics.Noop())
+	rMgr, err := bmanager.New(bmanager.NewResourceConfig(), r, log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	_ = manager.New(rMgr,
@@ -134,7 +134,7 @@ func TestTypeAPIDisabled(t *testing.T) {
 	assert.Greater(t, len(r.endpoints), 1)
 
 	r = &endpointReg{endpoints: map[string]http.HandlerFunc{}}
-	rMgr, err = bmanager.NewV2(bmanager.NewResourceConfig(), r, log.Noop(), metrics.Noop())
+	rMgr, err = bmanager.New(bmanager.NewResourceConfig(), r, log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	_ = manager.New(rMgr,
@@ -172,7 +172,7 @@ func harmlessConf() stream.Config {
 }
 
 func TestTypeAPIBasicOperations(t *testing.T) {
-	res, err := bmanager.NewV2(bmanager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
+	res, err := bmanager.New(bmanager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	mgr := manager.New(res)
@@ -285,7 +285,7 @@ func TestTypeAPIBasicOperations(t *testing.T) {
 }
 
 func TestTypeAPIPatch(t *testing.T) {
-	res, err := bmanager.NewV2(bmanager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
+	res, err := bmanager.New(bmanager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	mgr := manager.New(res)
@@ -337,7 +337,7 @@ func TestTypeAPIPatch(t *testing.T) {
 }
 
 func TestTypeAPIBasicOperationsYAML(t *testing.T) {
-	res, err := bmanager.NewV2(bmanager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
+	res, err := bmanager.New(bmanager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	mgr := manager.New(res)
@@ -414,7 +414,7 @@ func TestTypeAPIBasicOperationsYAML(t *testing.T) {
 }
 
 func TestTypeAPIList(t *testing.T) {
-	res, err := bmanager.NewV2(bmanager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
+	res, err := bmanager.New(bmanager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	mgr := manager.New(res)
@@ -449,7 +449,7 @@ func TestTypeAPIList(t *testing.T) {
 }
 
 func TestTypeAPISetStreams(t *testing.T) {
-	res, err := bmanager.NewV2(bmanager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
+	res, err := bmanager.New(bmanager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	mgr := manager.New(res)
@@ -521,7 +521,7 @@ func TestTypeAPISetStreams(t *testing.T) {
 }
 
 func TestTypeAPIStreamsDefaultConf(t *testing.T) {
-	res, err := bmanager.NewV2(bmanager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
+	res, err := bmanager.New(bmanager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	mgr := manager.New(res)
@@ -553,7 +553,7 @@ func TestTypeAPIStreamsDefaultConf(t *testing.T) {
 }
 
 func TestTypeAPIStreamsLinting(t *testing.T) {
-	res, err := bmanager.NewV2(bmanager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
+	res, err := bmanager.New(bmanager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	mgr := manager.New(res)
@@ -600,7 +600,7 @@ func TestTypeAPIStreamsLinting(t *testing.T) {
 }
 
 func TestTypeAPIDefaultConf(t *testing.T) {
-	res, err := bmanager.NewV2(bmanager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
+	res, err := bmanager.New(bmanager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	mgr := manager.New(res)
@@ -630,7 +630,7 @@ func TestTypeAPIDefaultConf(t *testing.T) {
 }
 
 func TestTypeAPILinting(t *testing.T) {
-	res, err := bmanager.NewV2(bmanager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
+	res, err := bmanager.New(bmanager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	mgr := manager.New(res)
@@ -730,7 +730,7 @@ func TestResourceAPILinting(t *testing.T) {
 
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			bmgr, err := bmanager.NewV2(bmanager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
+			bmgr, err := bmanager.New(bmanager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
 			require.NoError(t, err)
 
 			mgr := manager.New(bmgr)
@@ -767,7 +767,7 @@ func TestResourceAPILinting(t *testing.T) {
 }
 
 func TestTypeAPIGetStats(t *testing.T) {
-	mgr, err := bmanager.NewV2(bmanager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
+	mgr, err := bmanager.New(bmanager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	smgr := manager.New(mgr)
@@ -801,7 +801,7 @@ func TestTypeAPIGetStats(t *testing.T) {
 }
 
 func TestTypeAPISetResources(t *testing.T) {
-	bmgr, err := bmanager.NewV2(bmanager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
+	bmgr, err := bmanager.New(bmanager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	tChan := make(chan message.Transaction)