+++ public/service/plugins_test.go
@@ -61,7 +61,7 @@ test_cache_plugin_with_config:
 	require.NoError(t, err)
 	assert.Equal(t, cacheConfStr, string(cacheConfOutBytes))
 
-	mgr, err := manager.NewV2(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
+	mgr, err := manager.New(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	_, err = mgr.NewCache(cacheConf)
@@ -98,7 +98,7 @@ test_cache_plugin_without_config: null
 	require.NoError(t, err)
 	assert.Equal(t, cacheConfStr, string(cacheConfOutBytes))
 
-	mgr, err := manager.NewV2(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
+	mgr, err := manager.New(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	_, err = mgr.NewCache(cacheConf)
@@ -140,7 +140,7 @@ test_input_plugin_with_config:
 	require.NoError(t, err)
 	assert.Equal(t, inConfStr, string(outConfOutBytes))
 
-	mgr, err := manager.NewV2(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
+	mgr, err := manager.New(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	_, err = mgr.NewInput(inConf)
@@ -177,7 +177,7 @@ test_input_plugin_without_config: null
 	require.NoError(t, err)
 	assert.Equal(t, inConfStr, string(outConfOutBytes))
 
-	mgr, err := manager.NewV2(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
+	mgr, err := manager.New(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	_, err = mgr.NewInput(inConf)
@@ -218,7 +218,7 @@ test_output_plugin_with_config:
 	require.NoError(t, err)
 	assert.Equal(t, inConfStr, string(outConfOutBytes))
 
-	mgr, err := manager.NewV2(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
+	mgr, err := manager.New(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	_, err = mgr.NewOutput(inConf)
@@ -255,7 +255,7 @@ test_output_plugin_without_config: null
 	require.NoError(t, err)
 	assert.Equal(t, inConfStr, string(outConfOutBytes))
 
-	mgr, err := manager.NewV2(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
+	mgr, err := manager.New(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	_, err = mgr.NewOutput(inConf)
@@ -299,7 +299,7 @@ test_batch_output_plugin_with_config:
 	require.NoError(t, err)
 	assert.Equal(t, inConfStr, string(outConfOutBytes))
 
-	mgr, err := manager.NewV2(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
+	mgr, err := manager.New(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	_, err = mgr.NewOutput(inConf)
@@ -336,7 +336,7 @@ test_batch_output_plugin_without_config: null
 	require.NoError(t, err)
 	assert.Equal(t, inConfStr, string(outConfOutBytes))
 
-	mgr, err := manager.NewV2(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
+	mgr, err := manager.New(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	_, err = mgr.NewOutput(inConf)
@@ -376,7 +376,7 @@ test_processor_plugin_with_config:
 	require.NoError(t, err)
 	assert.Equal(t, inConfStr, string(outConfOutBytes))
 
-	mgr, err := manager.NewV2(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
+	mgr, err := manager.New(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	_, err = mgr.NewProcessor(inConf)
@@ -412,7 +412,7 @@ test_processor_plugin_without_config: null
 	require.NoError(t, err)
 	assert.Equal(t, inConfStr, string(outConfOutBytes))
 
-	mgr, err := manager.NewV2(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
+	mgr, err := manager.New(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	_, err = mgr.NewProcessor(inConf)
@@ -452,7 +452,7 @@ test_batch_processor_plugin_with_config:
 	require.NoError(t, err)
 	assert.Equal(t, inConfStr, string(outConfOutBytes))
 
-	mgr, err := manager.NewV2(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
+	mgr, err := manager.New(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	_, err = mgr.NewProcessor(inConf)
@@ -488,7 +488,7 @@ test_batch_processor_plugin_without_config: null
 	require.NoError(t, err)
 	assert.Equal(t, inConfStr, string(outConfOutBytes))
 
-	mgr, err := manager.NewV2(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
+	mgr, err := manager.New(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	_, err = mgr.NewProcessor(inConf)
@@ -528,7 +528,7 @@ test_rate_limit_plugin_with_config:
 	require.NoError(t, err)
 	assert.Equal(t, inConfStr, string(outConfOutBytes))
 
-	mgr, err := manager.NewV2(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
+	mgr, err := manager.New(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	_, err = mgr.NewRateLimit(inConf)
@@ -564,7 +564,7 @@ test_rate_limit_plugin_without_config: null
 	require.NoError(t, err)
 	assert.Equal(t, inConfStr, string(outConfOutBytes))
 
-	mgr, err := manager.NewV2(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
+	mgr, err := manager.New(manager.NewResourceConfig(), mock.NewManager(), log.Noop(), metrics.Noop())
 	require.NoError(t, err)
 
 	_, err = mgr.NewRateLimit(inConf)